Introduce current project
The current project is one made by TheOdinProject and I am about
to enter the next language which is CSS. I will be required to
make a website that has a homepage as well as multiple other pages
containing information on how to make dishes, they cna be either
ones that I know or can be provided. The main principle of this
project is to prove my knowledge and learnings of the topic of
HTML5. I hope this works out.